requirements.txt
pyyaml
inflector
matplotlib